Minor reorg of boilerplate:   App --> services/Services (with children)   Layout --> App   src/app/bootstrap --> /src   GameProvider --> DomainProvider   useGame --> useChess
