+++ apps/chess-web/index.html
@@ -17,6 +17,6 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-    <script type="module" src="src/app/bootstrap.tsx"></script> 
+    <script type="module" src="src/bootstrap.tsx"></script> 
   </body>
 </html>
