+++ apps/chess-web/src/app/widgets/SideMenu.tsx
@@ -4,7 +4,7 @@ import { observer } from 'mobx-react-lite'
 import type { GameSnapshot } from '@artemis-prime/chess-core'
 
 import { styled, type CSS } from '~/style'
-import { useChessboardOrientation, useGame, useSnapshotPersistence } from '~/services'
+import { useChessboardOrientation, useChess, useSnapshotPersistence } from '~/services'
 import { Drawer, type IconAndStyles } from '~/primatives'
 
 import {
@@ -49,7 +49,7 @@ const SideMenu: React.FC<{
 
   const bo = useChessboardOrientation()
   const sp = useSnapshotPersistence()
-  const game = useGame()
+  const game = useChess()
 
   const swapDirection = () => { bo.setWhiteOnBottom(!bo.whiteOnBottom) }
   const saveSnapshot = () => {
